add std container

+++ src/error.rs
@@ -6,7 +6,7 @@ pub type Result<T> = std::result::Result<T, Error>;
 #[snafu(visibility(pub(crate)))]
 pub enum Error {
     #[snafu(display("Data type mismatch, bits {size}, expect {expected}"))]
-    DataTypeSizeMismatch { size: usize, expected: u8 },
+    DataTypeSizeMismatch { size: usize, expected: usize },
 
     #[snafu(display("non contiguous"))]
     NonContiguous { shape: Vec<i64>, strides: Vec<i64> },
